Feature added and problem solved.
It’s added a form with the avalanches distribution and the sensitivity
charts (Charts provided by JFreeChart lib). It also possible analyze
the avalanches and sensitivity statistics (statistical features
provided by Common Math lib).
Now it is possible show the collapsed and not collapsed TES network (a
problem occurred with the columns names before).
The GESTODifferent style shows the nodes size in relation with the
normalised degree of the node. The normalised degree is calculated as a
normalisation of the node degree.